Testing Action run: Merging dev into stage #1

+++ database/migrations/2023_08_08_051827_create_imagefootage_order_items_table.php
@@ -20,10 +20,10 @@ class CreateImagefootageOrderItemsTable extends Migration
             $table->string('product_type')->nullable();
             $table->string('standard_type', 255)->nullable();
             $table->string('standard_size', 255)->nullable();
-            $table->integer('standard_price')->nullable();
+            $table->decimal('standard_price', 10, 2)->nullable();
             $table->string('extended_name', 255)->nullable();
-            $table->integer('extended_price')->nullable();
-            $table->integer('total')->nullable();
+            $table->decimal('extended_price', 10, 2)->nullable();
+            $table->decimal('total', 10, 2)->nullable();
             $table->string('product_name', 255)->nullable();
             $table->string('product_thumb', 255)->nullable();
             $table->text('product_desc')->nullable();
