+++ database/migrations/2023_08_08_051827_create_zip_codes_table.php
@@ -16,15 +16,15 @@ class CreateZipCodesTable extends Migration
         Schema::create('zip_codes', function (Blueprint $table) {
             $table->integer('id', true);
             $table->string('zip_code', 5)->index('zip_code');
-            $table->float('lat', 10, 0);
-            $table->float('lon', 10, 0);
+            $table->decimal('lat', 10, 2);
+            $table->decimal('lon', 10, 2);
             $table->string('city', 100)->index('city');
             $table->string('state_prefix', 100);
             $table->string('county', 100);
             $table->string('z_type', 100);
-            $table->float('xaxis', 10, 0);
-            $table->float('yaxis', 10, 0);
-            $table->float('zaxis', 10, 0);
+            $table->decimal('xaxis', 10, 2);
+            $table->decimal('yaxis', 10, 2);
+            $table->decimal('zaxis', 10, 2);
             $table->string('z_primary', 100);
             $table->string('country', 100)->index('country');
             $table->string('state_name', 255)->comment('US State Name');
